+++ src/Symfony/Component/DependencyInjection/Compiler/AutowireRequiredMethodsPass.php
@@ -36,6 +36,7 @@ class AutowireRequiredMethodsPass extends AbstractRecursivePass
 
         $alreadyCalledMethods = [];
         $withers = [];
+        $setters = [];
 
         foreach ($value->getMethodCalls() as [$method]) {
             $alreadyCalledMethods[strtolower($method)] = true;
@@ -49,11 +50,13 @@ class AutowireRequiredMethodsPass extends AbstractRecursivePass
             }
 
             while (true) {
-                if ($r->getAttributes(Required::class)) {
+                if ($attributes = $r->getAttributes(Required::class)) {
+                    $priority = $attributes[0]->newInstance()->priority ?? 0;
+
                     if ($this->isWither($r, $r->getDocComment() ?: '')) {
-                        $withers[] = [$r->name, [], true];
+                        $withers[] = [$r->name, [], $priority, true];
                     } else {
-                        $value->addMethodCall($r->name, []);
+                        $setters[] = [$r->name, [], $priority];
                     }
                     break;
                 }
@@ -64,11 +67,20 @@ class AutowireRequiredMethodsPass extends AbstractRecursivePass
             }
         }
 
+        usort($setters, static fn ($a, $b) => $b[2] <=> $a[2]);
+        foreach ($setters as $call) {
+            $value->addMethodCall($call[0], $call[1]);
+        }
+
         if ($withers) {
+            usort($withers, static fn ($a, $b) => $b[2] <=> $a[2]);
             // Prepend withers to prevent creating circular loops
-            $setters = $value->getMethodCalls();
-            $value->setMethodCalls($withers);
-            foreach ($setters as $call) {
+            $calls = $value->getMethodCalls();
+            $value->setMethodCalls([]);
+            foreach ($withers as $call) {
+                $value->addMethodCall($call[0], $call[1], $call[3]);
+            }
+            foreach ($calls as $call) {
                 $value->addMethodCall($call[0], $call[1], $call[2] ?? false);
             }
         }