+++ src/Symfony/Contracts/Service/Attribute/Required.php
@@ -22,4 +22,11 @@ namespace Symfony\Contracts\Service\Attribute;
 #[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY)]
 final class Required
 {
+    /**
+     * @param int $priority The priority of the method call when the class declares several required methods; the higher the sooner it's called
+     */
+    public function __construct(
+        public int $priority = 0,
+    ) {
+    }
 }